Marge Kavan commented validation into utils
